✨ feat: Add logging statements to parseNamespace function
This commit adds logging statements to the `parseNamespace` function in order to better understand the contents of the `namespace` variable. The logs output the values of each `ns` object in the `namespace` array. Additionally, the `getClasses` function is called for each `ns` object in order to log information about the `class` property.