✨ feat: Add Grid component to @react-gtk/core
The commit adds a new file `Grid.ts` to the `@react-gtk/core` package. This file contains the implementation of the `Grid` component, which extends the `Element` class. The `Grid` component is a wrapper around the Gtk.Grid from the Gtk library.

The `Grid` component provides a way to create grid layouts in React applications using Gtk. It accepts props similar to the native `Grid` component in JSX.IntrinsicElements and provides methods for appending and removing child widgets.

Additionally, the commit includes changes to the `generate` package's `index.tsx` file. It logs the `classes` property of the `Generator` instance to the console.