✨ feat: Add additional functionality to WidgetElement component
This commit adds additional functionality to the WidgetElement component in the generate package. The changes include importing various elements from react-ast, such as Class, ClassMethod, Import, Var, Code, ExportNamedDeclaration, ExportSpecifier, and Interface. The WidgetElementProps interface is also defined, with the addition of an optional extendedClass property. The WidgetElement function component now includes a default value for extendedClass if it is not provided. Inside the component, an interfaceName variable is created by concatenating the name prop with 'Props'. Some imports are added, including Element from the core package and Gtk from @girs/node-gtk-4.0. A new ClassMethod with the id "constructor" is added, which takes the props parameter and initializes a new Gtk.{name} instance. The super() method is called within the constructor. Finally, an ExportNamedDeclaration is added to export the {name} and {interfaceName} symbols.

Note: The diff also includes changes to the project-words.txt file, but they are unrelated to the commit message.