✨ feat: Add interface props to WidgetElement component
This commit adds support for interface props in the `WidgetElement` component. The `WidgetElementProps` interface now includes a new property `interfaceProps`, which is an array of `WidgetElementInterfaceProps`. The `WidgetElement` component has been updated to render the interface props as `PropertySignature` elements within the component. The interface props are mapped over and rendered with their respective name and type annotations.

In addition, the `Generator` class has been modified to generate default interface props for the `WidgetElement` component. These default props include `children` of type `ReactNode`, `ref` of type `Ref<PublicInstance<Gtk.Widget>>`, and `sizeRequest` of type `[number, number]`.

The `renderWidgetElement` function now accepts `interfaceProps` as an option and passes it down to the `WidgetElement` component during rendering.